examples/c: Add parse_args to common & make use for phi samples
